+++ src/bootstrap-table.js
@@ -2510,6 +2510,7 @@
         });
 
         this.initSearch();
+        this.initPagination();
         this.initSort();
         this.initBody(true);
     };